Creates separate folder for booking components
