+++ src/App.js
@@ -16,8 +16,8 @@ import PasswordReset from './Components/Authentication/PasswordReset'
 import ChangePassword from './Components/Authentication/ChangePassword'
 import PasswordResetSuccess from './Components/Authentication/PasswordResetSuccess'
 import UserPage from './Components/UserPage/UserPage'
-import RequestToBook from './Components/RequestToBook'
-import SuccessfulRequest from './Components/SuccessfulRequest'
+import RequestToBook from './Components/Bookings/RequestToBook'
+import SuccessfulRequest from './Components/Bookings/SuccessfulRequest'
 import ScrollToTop from './Modules/ScrollToTop'
 import { Container, Sticky, Sidebar } from 'semantic-ui-react'
 import { connect } from 'react-redux'
