+++ src/Components/Bookings/RequestToBook.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux'
 import { Header, Form, Button, Message, Segment } from 'semantic-ui-react'
 import moment from 'moment'
 import axios from 'axios'
-import { pricePerDay, total } from '../Modules/PriceCalculations'
+import { pricePerDay, total } from '../../Modules/PriceCalculations'
 
 class RequestToBook extends Component {
